refactor: cleanup schemas
- combine account recovery schemas into account recovery router

- remove unused schemas/types
